requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.4

# CLI enhancements
click==8.1.7
rich==13.7.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1

# Code quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
types-requests==2.31.0.10

# Utilities
tenacity==8.2.3  # For retry logic
validators==0.22.0  # For email validation